pyproject.toml
[project]
name = "zestify"
version = "0.1.0"
description = "AI-powered personal wellness and fitness assistant"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.7",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[project.scripts]
zestify = "backend.services.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]
only-include = ["backend"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.black]
line-length = 88
target-version = ['py311']

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"